Wellspring/files/thread-3/pyproject.toml
[project]
name = "wellspring-thread3"
version = "0.1.0"
description = "WoT Thread 3: Pool Control + Execution Loop"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pynacl>=1.5.0",
    "blake3>=0.4.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
agent = [
    "anthropic>=0.20.0",
]
rag = [
    "sentence-transformers>=2.2.0",
]
all = [
    "anthropic>=0.20.0",
    "sentence-transformers>=2.2.0",
]

[project.scripts]
wot-daemon = "daemon:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
